@@ -1,6 +1,6 @@
 {
   "name": "@ampproject/remapping",
-  "version": "1.1.1",
+  "version": "2.2.0",
   "description": "Remap sequential sourcemaps through transformations to point at the original source code",
   "keywords": [
     "source",
@@ -32,7 +32,7 @@
     "prebuild": "rm -rf dist",
     "prepublishOnly": "npm run preversion",
     "preversion": "run-s test build",
-    "test": "run-s -n test:lint 'test:only -- --no-cache'",
+    "test": "run-s -n test:lint test:only",
     "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand",
     "test:lint": "run-s -n test:lint:*",
     "test:lint:prettier": "prettier --check '{src,test}/**/*.ts'",
@@ -41,23 +41,23 @@
     "test:watch": "jest --coverage --watch"
   },
   "devDependencies": {
-    "@rollup/plugin-typescript": "8.1.0",
-    "@types/jest": "26.0.19",
-    "@typescript-eslint/eslint-plugin": "4.10.0",
-    "@typescript-eslint/parser": "4.10.0",
-    "eslint": "7.15.0",
-    "eslint-config-prettier": "7.0.0",
-    "jest": "26.6.3",
-    "jest-config": "26.6.3",
+    "@rollup/plugin-typescript": "8.3.2",
+    "@types/jest": "27.4.1",
+    "@typescript-eslint/eslint-plugin": "5.20.0",
+    "@typescript-eslint/parser": "5.20.0",
+    "eslint": "8.14.0",
+    "eslint-config-prettier": "8.5.0",
+    "jest": "27.5.1",
+    "jest-config": "27.5.1",
     "npm-run-all": "4.1.5",
-    "prettier": "2.2.1",
-    "rollup": "2.35.1",
-    "ts-jest": "26.4.4",
-    "tslib": "2.0.3",
-    "typescript": "4.1.3"
+    "prettier": "2.6.2",
+    "rollup": "2.70.2",
+    "ts-jest": "27.1.4",
+    "tslib": "2.4.0",
+    "typescript": "4.6.3"
   },
   "dependencies": {
-    "@jridgewell/resolve-uri": "^3.0.3",
-    "sourcemap-codec": "1.4.8"
+    "@jridgewell/gen-mapping": "^0.1.0",
+    "@jridgewell/trace-mapping": "^0.3.9"
   }
 }
