@@ -1,22 +1,19 @@
 {
   "name": "@angular/cli",
-  "version": "13.2.4",
+  "version": "15.1.5",
   "description": "CLI tool for Angular",
   "main": "lib/cli/index.js",
   "bin": {
     "ng": "./bin/ng.js"
   },
   "keywords": [
-    "angular",
     "Angular CLI",
-    "devkit",
-    "sdk",
     "Angular DevKit",
-    "angular-cli"
+    "angular",
+    "angular-cli",
+    "devkit",
+    "sdk"
   ],
-  "scripts": {
-    "postinstall": "node ./bin/postinstall/script.js"
-  },
   "repository": {
     "type": "git",
     "url": "https://github.com/angular/angular-cli.git"
@@ -28,39 +25,38 @@
   },
   "homepage": "https://github.com/angular/angular-cli",
   "dependencies": {
-    "@angular-devkit/architect": "0.1302.4",
-    "@angular-devkit/core": "13.2.4",
-    "@angular-devkit/schematics": "13.2.4",
-    "@schematics/angular": "13.2.4",
+    "@angular-devkit/architect": "0.1501.5",
+    "@angular-devkit/core": "15.1.5",
+    "@angular-devkit/schematics": "15.1.5",
+    "@schematics/angular": "15.1.5",
     "@yarnpkg/lockfile": "1.1.0",
-    "ansi-colors": "4.1.1",
-    "debug": "4.3.3",
-    "ini": "2.0.0",
-    "inquirer": "8.2.0",
-    "jsonc-parser": "3.0.0",
-    "npm-package-arg": "8.1.5",
-    "npm-pick-manifest": "6.1.1",
+    "ansi-colors": "4.1.3",
+    "ini": "3.0.1",
+    "inquirer": "8.2.4",
+    "jsonc-parser": "3.2.0",
+    "npm-package-arg": "10.1.0",
+    "npm-pick-manifest": "8.0.1",
     "open": "8.4.0",
     "ora": "5.4.1",
-    "pacote": "12.0.3",
-    "resolve": "1.22.0",
-    "semver": "7.3.5",
+    "pacote": "15.0.8",
+    "resolve": "1.22.1",
+    "semver": "7.3.8",
     "symbol-observable": "4.0.0",
-    "uuid": "8.3.2"
+    "yargs": "17.6.2"
   },
   "ng-update": {
     "migrations": "@schematics/angular/migrations/migration-collection.json",
     "packageGroup": {
-      "@angular/cli": "13.2.4",
-      "@angular-devkit/architect": "0.1302.4",
-      "@angular-devkit/build-angular": "13.2.4",
-      "@angular-devkit/build-webpack": "0.1302.4",
-      "@angular-devkit/core": "13.2.4",
-      "@angular-devkit/schematics": "13.2.4"
+      "@angular/cli": "15.1.5",
+      "@angular-devkit/architect": "0.1501.5",
+      "@angular-devkit/build-angular": "15.1.5",
+      "@angular-devkit/build-webpack": "0.1501.5",
+      "@angular-devkit/core": "15.1.5",
+      "@angular-devkit/schematics": "15.1.5"
     }
   },
   "engines": {
-    "node": "^12.20.0 || ^14.15.0 || >=16.10.0",
+    "node": "^14.20.0 || ^16.13.0 || >=18.10.0",
     "npm": "^6.11.0 || ^7.5.6 || >=8.0.0",
     "yarn": ">= 1.13.0"
   }
