@@ -1,6 +1,6 @@
 {
   "name": "@angular/compiler-cli",
-  "version": "13.1.3",
+  "version": "15.1.4",
   "description": "Angular - the compiler CLI for Node.js",
   "typings": "index.d.ts",
   "bin": {
@@ -53,21 +53,20 @@
     }
   },
   "dependencies": {
-    "@babel/core": "^7.8.6",
+    "@babel/core": "7.19.3",
+    "@jridgewell/sourcemap-codec": "^1.4.14",
     "reflect-metadata": "^0.1.2",
-    "canonical-path": "1.0.0",
     "chokidar": "^3.0.0",
     "convert-source-map": "^1.5.1",
     "dependency-graph": "^0.11.0",
-    "magic-string": "^0.25.0",
+    "magic-string": "^0.27.0",
     "semver": "^7.0.0",
-    "sourcemap-codec": "^1.4.8",
     "tslib": "^2.3.0",
     "yargs": "^17.2.1"
   },
   "peerDependencies": {
-    "@angular/compiler": "13.1.3",
-    "typescript": ">=4.4.2 <4.6"
+    "@angular/compiler": "15.1.4",
+    "typescript": ">=4.8.2 <5.0"
   },
   "repository": {
     "type": "git",
@@ -80,12 +79,12 @@
   ],
   "license": "MIT",
   "engines": {
-    "node": "^12.20.0 || ^14.15.0 || >=16.10.0"
+    "node": "^14.20.0 || ^16.13.0 || >=18.10.0"
   },
   "bugs": {
     "url": "https://github.com/angular/angular/issues"
   },
-  "homepage": "https://github.com/angular/angular/tree/master/packages/compiler-cli",
+  "homepage": "https://github.com/angular/angular/tree/main/packages/compiler-cli",
   "ng-update": {
     "packageGroup": [
       "@angular/core",
