@@ -1,14 +1,22 @@
 {
   "name": "@angular/compiler",
-  "version": "13.1.3",
+  "version": "15.1.4",
   "description": "Angular - the compiler library",
   "author": "angular",
   "license": "MIT",
   "engines": {
-    "node": "^12.20.0 || ^14.15.0 || >=16.10.0"
+    "node": "^14.20.0 || ^16.13.0 || >=18.10.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
+  },
+  "peerDependencies": {
+    "@angular/core": "15.1.4"
+  },
+  "peerDependenciesMeta": {
+    "@angular/core": {
+      "optional": true
+    }
   },
   "repository": {
     "type": "git",
@@ -56,7 +64,7 @@
       "default": "./fesm2020/compiler.mjs"
     },
     "./testing": {
-      "types": "./testing/testing.d.ts",
+      "types": "./testing/index.d.ts",
       "esm2020": "./esm2020/testing/testing.mjs",
       "es2020": "./fesm2020/testing.mjs",
       "es2015": "./fesm2015/testing.mjs",
