@@ -1,11 +1,11 @@
 {
   "name": "@angular/core",
-  "version": "13.1.3",
+  "version": "15.1.4",
   "description": "Angular - the core framework",
   "author": "angular",
   "license": "MIT",
   "engines": {
-    "node": "^12.20.0 || ^14.15.0 || >=16.10.0"
+    "node": "^14.20.0 || ^16.13.0 || >=18.10.0"
   },
   "exports": {
     "./schematics/*": {
@@ -15,7 +15,7 @@
       "default": "./package.json"
     },
     ".": {
-      "types": "./core.d.ts",
+      "types": "./index.d.ts",
       "esm2020": "./esm2020/core.mjs",
       "es2020": "./fesm2020/core.mjs",
       "es2015": "./fesm2015/core.mjs",
@@ -23,7 +23,7 @@
       "default": "./fesm2020/core.mjs"
     },
     "./testing": {
-      "types": "./testing/testing.d.ts",
+      "types": "./testing/index.d.ts",
       "esm2020": "./esm2020/testing/testing.mjs",
       "es2020": "./fesm2020/testing.mjs",
       "es2015": "./fesm2015/testing.mjs",
@@ -36,7 +36,7 @@
   },
   "peerDependencies": {
     "rxjs": "^6.5.3 || ^7.4.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.11.4 || ~0.12.0"
   },
   "repository": {
     "type": "git",
@@ -68,7 +68,7 @@
   "fesm2020": "./fesm2020/core.mjs",
   "fesm2015": "./fesm2015/core.mjs",
   "esm2020": "./esm2020/core.mjs",
-  "typings": "./core.d.ts",
+  "typings": "./index.d.ts",
   "module": "./fesm2015/core.mjs",
   "es2020": "./fesm2020/core.mjs",
   "type": "module"
