@@ -1,19 +1,19 @@
 {
   "name": "@angular/forms",
-  "version": "13.1.3",
+  "version": "15.1.4",
   "description": "Angular - directives and services for creating forms",
   "author": "angular",
   "license": "MIT",
   "engines": {
-    "node": "^12.20.0 || ^14.15.0 || >=16.10.0"
+    "node": "^14.20.0 || ^16.13.0 || >=18.10.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/core": "13.1.3",
-    "@angular/common": "13.1.3",
-    "@angular/platform-browser": "13.1.3",
+    "@angular/core": "15.1.4",
+    "@angular/common": "15.1.4",
+    "@angular/platform-browser": "15.1.4",
     "rxjs": "^6.5.3 || ^7.4.0"
   },
   "repository": {
@@ -45,7 +45,7 @@
   "fesm2020": "./fesm2020/forms.mjs",
   "fesm2015": "./fesm2015/forms.mjs",
   "esm2020": "./esm2020/forms.mjs",
-  "typings": "./forms.d.ts",
+  "typings": "./index.d.ts",
   "module": "./fesm2015/forms.mjs",
   "es2020": "./fesm2020/forms.mjs",
   "type": "module",
@@ -54,7 +54,7 @@
       "default": "./package.json"
     },
     ".": {
-      "types": "./forms.d.ts",
+      "types": "./index.d.ts",
       "esm2020": "./esm2020/forms.mjs",
       "es2020": "./fesm2020/forms.mjs",
       "es2015": "./fesm2015/forms.mjs",
