@@ -1,10 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { ViewChild } from '@angular/core';
 import { ElementRef } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { SharedDataService } from 'src/app/services/shared-data.service';
 import { Router } from '@angular/router';
+import { SocialAuthService, SocialUser, GoogleLoginProvider } from '@abacritt/angularx-social-login';
 
 @Component({
   selector: 'form-register',
@@ -12,7 +13,10 @@ import { Router } from '@angular/router';
   styleUrls: ['./form-register.component.css']
 })
 export class FormRegisterComponent implements OnInit {
-  registerForm!: FormGroup;
+  user:SocialUser;
+  loggedIn:boolean;
+  accessToken:string;
+  registerForm!: UntypedFormGroup;
   passW!: string;
   confirmPass!: string;
   isPass: boolean = true;
@@ -21,10 +25,37 @@ export class FormRegisterComponent implements OnInit {
   cpass!: string;
   isCshow!: boolean;
   fakeUrl: string = 'http://localhost:4200/';
-  constructor(private formBuilder:FormBuilder, private http:HttpClient, public sharedDataService:SharedDataService, private router:Router) { }
+  constructor(private formBuilder:UntypedFormBuilder, private http:HttpClient, public sharedDataService:SharedDataService, 
+    private router:Router, private authService:SocialAuthService) {
+    }
 
 
   ngOnInit(): void {
+      this.authService.authState.subscribe((user: SocialUser) => {
+        this.user = user;
+        this.loggedIn = (user!=null);
+        if(this.loggedIn) {
+          this.http.get("https://backend-dot-lightscreendotart.uk.r.appspot.com/signupWithExternal?idtoken="+this.user.idToken+"&provider="+this.user.provider).subscribe(result => {
+            this.sharedDataService.userInfo = JSON.stringify(result).split('[').join("").split(']').join("").split('"').join("").split(","); 
+            if(this.sharedDataService.userInfo.length > 1) {
+              this.sharedDataService.signedIn = true;
+              if((<HTMLInputElement>document.getElementById("rememberMeBox"))?.checked) {localStorage.setItem('userInfo', JSON.stringify(this.sharedDataService.userInfo));}
+              this.router.navigate(['/']);
+            }
+            else if(this.sharedDataService.userInfo.length == 1 && this.sharedDataService.userInfo[0] == -1) {
+              // alert("A user with that email already exists.");
+              this.http.get("https://backend-dot-lightscreendotart.uk.r.appspot.com/loginWithExternal?idtoken="+this.user.idToken+"&provider="+this.user.provider).subscribe(result => {
+                this.sharedDataService.userInfo = JSON.stringify(result).split('[').join("").split(']').join("").split('"').join("").split(","); 
+                if(this.sharedDataService.userInfo.length > 1) {
+                  this.sharedDataService.signedIn = true;
+                  if((<HTMLInputElement>document.getElementById("rememberMeBox"))?.checked) {localStorage.setItem('userInfo', JSON.stringify(this.sharedDataService.userInfo));}
+                  this.router.navigate(['/']);
+                } 
+              });
+            } 
+          });
+        }
+      });
     this.registerForm = this.formBuilder.group({
       emailId: ['', [Validators.required, Validators.email]],
       password: ['', [Validators.required, Validators.minLength(8), Validators.maxLength(24)]],
